feat(input): use password component if `Input` has `type=password`
We want to provide unified experience to enter a password. To avoid the inaccessible and manually handled password input, we use `Password` component to render `Input` with `type=password`.

We keep an escape hatch in the form of `InnerInput` component if user needs a default html password field
